_.max and _.min map arrays correctly [fix #2037]
Currently, mapping over a nested array and using _.max or _.min produces unexpected results.
![infinity](https://cloud.githubusercontent.com/assets/5640348/6101277/3c2170d4-afd6-11e4-8b25-cd6e6a0d4276.png)

This occurs because when _.max or _.min is passed into _.map, the function optimizeCb attempts to account for the fact that _.max and _.min have multiple arguments by trying to pass an index value into them. This creates a misalignment between the parameters being passed into _.max and _.min and what they are actually expecting. _.max and _.min expect for an iteratee to be passed in as a second argument, but instead _.map is passing an index.

The result of this arguments mismatch makes _.min and _.max fail the following conditional which checks to see if an iteratee was passed.
![screen shot 2015-02-08 at 9 10 57 pm](https://cloud.githubusercontent.com/assets/5640348/6101333/3a8ada48-afd7-11e4-90e1-253685872050.png)

I have corrected the above conditional by including a following check to ignore the iteratee if it is a number:
![works](https://cloud.githubusercontent.com/assets/5640348/6101347/90681da4-afd7-11e4-9fb3-5a4964c3f668.png)

This additional conditional makes _.min and _.max produce the expected results with no side effects to their normal behavior.
Nothing else has been changed and tests have been added to verify the code I have added is indeed valid and does not disrupt
anything.